+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 #TBD
-* Add Native Auth V2 email OTP sign-in.
+* Add Native Auth V2 email OTP sign-in #3085
 * Add SignIn v2 #3079
 
 ## [2.15.0]